test: move repeated params kernel to fixtures
RepeatedDefaultParametersTest was the only file in tests/Functional
without a namespace, breaking the ApiPlatform\Tests\ PSR-4 rule, and
declared its kernel inline instead of tests/Fixtures/app/.
